tests/submission/PhilipTaylor/canvas/initial.reset.different.html

<!DOCTYPE html>
<html><head><meta charset="utf-8">
<title>Canvas test: initial.reset.different</title>
<meta content="Philip Taylor" name="author">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../common/canvas-tests.js"></script>
<link href="../common/canvas-tests.css" rel="stylesheet">
</head><body class="show_output">

<h1><a href="index.initial.html">initial</a>.<a href="index.initial.reset.html">reset</a>.different</h1>
<p class="desc">Changing size resets canvas to transparent black</p>

<div class="refs">Spec references:
<ul>
<li><a href="../annotated-spec/canvas.html#testrefs.initial.reset">initial.reset</a>
</li></ul>
</div>

<p class="output">Actual output:</p>
<canvas class="output" height="50" id="c" width="100"><p class="fallback">FAIL (fallback content)</p></canvas>
<p class="output expectedtext">Expected output:</p><p><img alt="" class="output expected" id="expected" src="initial.reset.different.png">
</p><ul id="d"></ul>
<script>
var t = async_test("Changing size resets canvas to transparent black");
_addTest(function(canvas, ctx) {

ctx.fillStyle = '#f00';
ctx.fillRect(0, 0, 50, 50);
_assertPixel(canvas, 20,20, 255,0,0,255, "20,20", "255,0,0,255");
canvas.width = 50;
_assertPixel(canvas, 20,20, 0,0,0,0, "20,20", "0,0,0,0");


});
</script>

<div id="log"></div>
</body></html>
